General cleanup & fixup (#4)
* Remove bogus "src" entries from .gitignore files

* Sync all package .gitignore files with flutter/plugins/.gitignore

https://github.com/flutter/plugins/blob/master/.gitignore

* Cleanup generated/ignored files

Everything matching .gitignore (git ls-files -i --exclude-standard).

* Cleanup bogus ios-folders in non-iOS packages

* Remove connectivity_macos/example

First of all, connectivity_macos doesn't need its own example, but the
shared example in the main connectivity_plus package will do. Secondly,
the rest of connectivity_macos was moved to connectivity_plus_macos.

* share_plus: restore files that went missing in the mono-repo transition

Probably because "src" was listed in .gitignore.

* Fix podspecs for macOS & iOS

* Fix FLTPackageInfoPlugin.m on iOS

Apparently it was a symbolic link before, but it got screwed up in the
transition. The macOS version is now in a separate repo, so there's no
reason to have a shared folder for Darwin in the main connectivity pkg.